Refector debug info parser into debug info
